docs:コードの移行ミスを修正
