+++ src/model.ts
@@ -1,12 +1,55 @@
-export function checkNullSquareExist(Board: string[][]): boolean {
-  let isNullSquareExist = false;
-  Board.forEach((nthRow) => {
-    nthRow.forEach((nthSquare) => {
-      if (!nthSquare) {
-        isNullSquareExist = true;
-      }
-    });
-  });
-
-  return isNullSquareExist;
+export function calculateWinner(board: string[][]): number[][] | null {
+  const lines = [
+    [
+      [0, 0],
+      [0, 1],
+      [0, 2],
+    ],
+    [
+      [1, 0],
+      [1, 1],
+      [1, 2],
+    ],
+    [
+      [2, 0],
+      [2, 1],
+      [2, 2],
+    ],
+    [
+      [0, 0],
+      [1, 0],
+      [2, 0],
+    ],
+    [
+      [0, 1],
+      [1, 1],
+      [2, 1],
+    ],
+    [
+      [0, 2],
+      [1, 2],
+      [2, 2],
+    ],
+    [
+      [0, 0],
+      [1, 1],
+      [2, 2],
+    ],
+    [
+      [0, 2],
+      [1, 1],
+      [2, 0],
+    ],
+  ];
+  for (let i = 0; i < lines.length; i++) {
+    const [[a, x], [b, y], [c, z]] = lines[i];
+    if (
+      board[a][x] &&
+      board[a][x] === board[b][y] &&
+      board[a][x] === board[c][z]
+    ) {
+      return lines[i];
+    }
+  }
+  return null;
 }
